fixing some edits and moving work related to registration and authorization
