+++ src/main/java/ru/daniil/NauJava/controller/AuthController.java
@@ -1,6 +1,5 @@
 package ru.daniil.NauJava.controller;
 
-import io.swagger.v3.oas.annotations.Operation;
 import jakarta.validation.Valid;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -11,13 +10,13 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import ru.daniil.NauJava.request.RegistrationRequest;
 import ru.daniil.NauJava.service.UserDetailsServiceImpl;
-import ru.daniil.NauJava.service.UserService;
+import ru.daniil.NauJava.service.UserServiceImpl;
 
 @Controller
 public class AuthController {
-    private final UserDetailsServiceImpl userService;
+    private final UserServiceImpl userService;
 
-    public AuthController(UserDetailsServiceImpl userService) {
+    public AuthController(UserServiceImpl userService) {
         this.userService = userService;
     }
