+++ src/main/java/ru/daniil/NauJava/controller/NewExceptionHandler.java
@@ -19,7 +19,8 @@ public class NewExceptionHandler {
     /**
      * Данный метод обрабатывает ошибку о не найденном ресурсе (странице или данных), а после
      * создаёт объект класса ApiError и возвращает страницу с ошибкой
-     * @param ex ошибка
+     *
+     * @param ex      ошибка
      * @param request запрос в котором произошла ошибка
      * @return страница с ошибкой
      */
@@ -38,7 +39,8 @@ public class NewExceptionHandler {
     /**
      * Данный метод обрабатывает ошибку невалидных данных в одном поле, а после
      * создаёт объект класса ApiError и возвращает страницу с ошибкой
-     * @param ex ошибка
+     *
+     * @param ex      ошибка
      * @param request запрос в котором произошла ошибка
      * @return страница с ошибкой
      */
@@ -57,7 +59,8 @@ public class NewExceptionHandler {
     /**
      * Данный метод обрабатывает ошибку невалидных данных во множестве полей, а после
      * создаёт объект класса ApiError и возвращает страницу с ошибкой
-     * @param ex ошибка
+     *
+     * @param ex      ошибка
      * @param request запрос в котором произошла ошибка
      * @return страница с ошибкой
      */
