+++ src/main/java/ru/daniil/NauJava/repository/UserRepository.java
@@ -26,7 +26,5 @@ public interface UserRepository extends CrudRepository<User, Long> {
 
     boolean existsByLogin(String login);
 
-    boolean existsByLoginAndPassword(String login, String password);
-
     boolean existsByEmail(String email);
 }
